refactor: move inner function to seperate file
