+++ src/jvmMain/kotlin/ui/fileview/IconMode.kt
@@ -1,0 +1,8 @@
+package ui.fileview
+
+enum class IconMode {
+    Open,
+    Closed,
+    Selected,
+    Unselected,
+}
