+++ src/jvmMain/kotlin/ui/fileview/Node.kt
@@ -1,0 +1,129 @@
+package ui.fileview
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.TooltipArea
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.onClick
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.rounded.KeyboardArrowDown
+import androidx.compose.material.icons.rounded.KeyboardArrowRight
+import androidx.compose.material.icons.rounded.RadioButtonChecked
+import androidx.compose.material.icons.rounded.RadioButtonUnchecked
+import androidx.compose.material.ripple.rememberRipple
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextIndent
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+/**
+ * The state of a node in the tree view needed to render it.
+ */
+class NodeState(
+    val content: String,
+    val indentation: Dp,
+    val iconMode: IconMode,
+    val modifier: Modifier = Modifier,
+    val closeCallback: (() -> Unit)? = null,
+    val onClick: () -> Unit,
+)
+
+/**
+ * A node in the tree view.
+ * This is the basic building block of the tree view, and the UI for a single node.
+ */
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun Node(
+    state: NodeState,
+) {
+    val backgroundColor = if (state.iconMode == IconMode.Selected) {
+        MaterialTheme.colorScheme.outlineVariant
+    } else {
+        MaterialTheme.colorScheme.surface
+    }
+
+    Row(
+        state.modifier.padding(start = state.indentation).background(backgroundColor, shape = MaterialTheme.shapes.small)
+            .padding(start = 4.dp, top = 4.dp, bottom = 4.dp, end = 8.dp).onClick(onClick = state.onClick),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        when (state.iconMode) {
+            IconMode.Open -> Icon(Icons.Rounded.KeyboardArrowDown, contentDescription = "Drawer handle is open")
+            IconMode.Closed -> Icon(
+                Icons.Rounded.KeyboardArrowRight,
+                contentDescription = "Drawer handle is closed",
+            )
+
+            IconMode.Selected -> Icon(
+                Icons.Rounded.RadioButtonChecked,
+                contentDescription = "Icon checked",
+                Modifier.scale(0.5F),
+                tint = MaterialTheme.colorScheme.primary,
+            )
+
+            IconMode.Unselected -> Icon(
+                Icons.Rounded.RadioButtonUnchecked,
+                contentDescription = "Icon unchecked",
+                Modifier.scale(0.5F),
+            )
+        }
+
+        TooltipArea(tooltip = {
+            Surface(
+                modifier = Modifier.shadow(4.dp),
+                color = MaterialTheme.colorScheme.surfaceVariant,
+                shape = MaterialTheme.shapes.extraSmall,
+            ) {
+                Text(
+                    text = state.content,
+                    modifier = Modifier.padding(4.dp),
+                    style = MaterialTheme.typography.labelMedium,
+                )
+            }
+        }) {
+            Text(
+                state.content,
+                style = TextStyle(textIndent = TextIndent(0.sp, 12.sp)),
+            )
+        }
+
+        if (state.closeCallback != null) {
+            // An IconButton is currently fixed to 48.dp, so we need to make our own.
+            // https://github.com/androidx/androidx/blob/androidx-main/compose/material/material/src/commonMain/kotlin/androidx/compose/material/IconButton.kt
+            Box(
+                modifier = Modifier.size(16.dp).padding(start = 4.dp)
+                    .clickable(
+                        role = Role.Button,
+                        interactionSource = MutableInteractionSource(),
+                        indication = rememberRipple(bounded = false, radius = 12.dp),
+                        onClick = state.closeCallback,
+                    ),
+                contentAlignment = Alignment.Center,
+            ) {
+                Icon(
+                    Icons.Default.Close,
+                    tint = MaterialTheme.colorScheme.error,
+                    contentDescription = "Close file",
+                )
+            }
+        }
+    }
+}
